Data analysis & data visualisation
Added analysis_functions.py to calculate the minimum and maximum numbers in a list.

Added get_numbers to extraction.py to extract all the dollar amounts from a text.

Added data-analysis.py to implement the above and make it into a .csv.

Added graphical-analysis.py to plot the minimums and maximums on a histogram rounded to the nearest 1000s.